== SoundSys (OpenAL implementation) == What was intended as a minor revision of BufferT code turned into a good overhaul:
The original (and only) motivation was the removal of method IsStream() from the BufferT hierarchy.
This implied a revision of the interaction/cooperation between the buffers and mixer tracks,
especially regarding the automatic restarting of playback should buffer queues run empty.

The first but failed attempt was to use the "buffer looping" feature of OpenAL, see
  * thread <http://opensource.creative.com/pipermail/openal/2010-June/012128.html> and
  * the related code that has been left in but commented out with //DELETE
for details.

It turned out that the proper solution is to consider (e.g. the streaming) buffers as state machines
that have (implicit) states "initialized" (but not yet playing), "playing", and "stopped".
See the comment in file StreamingBuffer.cpp beginning in line 134 for more details.

Structurally, the most important changes are that BufferT::Update() is now called independently from
the state of the mixer track (OpenAL source), and that stopping playback now immediately releases the
whole "sound - mixer track - buffer" relationship. Attaching and detaching buffers to mixer tracks is
now a crucial part of the initialization for playing a sound.

These changes result in simpler and more robust code and meet the original intents for this revision.

== General ==
  * Updated the SConstruct file for compatibility with SCons 2.0.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@85 5fe0b3f3-48df-4727-b460-e6dda9226989