CaWE: Moved the code for activating the 3D view camera control from the Camera tool into the 3D view itself.
As the RMB was chosen as the main means for activating the 3D camera control, it's handling has been revised to allow for its dual function of opening the context menu or the current tools right-click action as well.

As a result, the 3D view can now be navigated independently from the currently active tool: The user can assume camera control at any time by pressing the relevant buttons (RMB or SPACE, or Z for toggle) without being required to first switch to the Camera tool.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@167 5fe0b3f3-48df-4727-b460-e6dda9226989