Added SConscript files for new external libraries expat and pcre.
For expat, the public interface headers were moved into the main directory, so that applications don't have to search the lib/ directory and then by accident find headers they didn't mean to find.

For pcre, additional files that are required for building were created according to the documentation in NON-UNIX-USE. The same key idea of separating public and internal headers as for expat was applied, but this time the public headers are in the include/ subdirectory.

These changes were made under Windows, related adjustments for Linux are scheduled for the next revision.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@226 5fe0b3f3-48df-4727-b460-e6dda9226989