vendor/OpenCollada: Initial import of OpenCollada at r788 from <http://code.google.com/p/opencollada/>, with the patch by daChaac for issue 120 applied (VC++ 2010 compilation, <http://code.google.com/p/opencollada/issues/detail?id=120#c5>).
Only the required subset of the files has been imported, in order to help clarity and to conserve disk-space and bandwidth.
The required set of files has been determined in advance by implementing a Collada stub (test) loader that is committed in one of the next revisions.

For more about OpenCollada, see also:
<http://www.opencollada.org>
<http://wiki.blender.org/index.php/User:Chingachgook>


git-svn-id: https://srv7.svn-repos.de/dev123/projects/vendor@231 5fe0b3f3-48df-4727-b460-e6dda9226989