Cafu models: Updated/fixed all loaders to create proper and reasonable default materials whenever a prior definition for the material was not found in the models material manager / .cmat file. (Note that .cmat files are mandatory only for .cmdl models, and an entirely optional extra for all others.)
The created default materials are typically fail-safe wire-frame materials. Alternatively, when more exhaustive data (like texture images) is available, an effort is made to create a material that is as close as possible to the original. The latter applies especially to HL1 mdl files and all file formats loaded by the FBX loader (.fbx, .dae, .3ds, .obj, .dxf).

This revision continues/completes the work begun r287.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@290 5fe0b3f3-48df-4727-b460-e6dda9226989