Cafu model code and Model Editor: Added full support for submodels.
The new code can now easily and very elegantly handle "super" models whose skeleton pose should be used when rendering the current model.
For example, a player model can act as the super model for a weapon, so that the skeleton of the weapon is copied from the player model in order to align the weapon with the hands of the player.

It is easily possible to load '''multiple''' submodels for any given super model, e.g. independent weapons for the left and right hands of a player model.

The Model Editor has been extended accordingly to allow loading and unloading of submodels for the currently loaded main ("super") model, and to render them together.

Also updated the Model Editor to use `EVT_PG_CHANGING` rather than `EVT_PG_CHANGED` in the joint inspector.

Closes #75.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@332 5fe0b3f3-48df-4727-b460-e6dda9226989