GuiSys and Gui Editor: This big change much improves the code design, separating the core GuiSys and its window classes from the Gui Editor more clearly, so that both are easier to understand and easier to expand or change without one affecting the other.
The key idea is to keep the `cf::GuiSys::WindowT` class hierarchy entirely unaware of the Gui Editor: All they ever see is an abstract `cf::GuiSys::WindowT::ExtDataT` instance that user code can derive from and assign to a GuiSys window in order to "configure" it with "callbacks" (e.g. the `Render()` method) and to have it store additional user-specific data and functions.

The Gui Editor derives from `cf::GuiSys::WindowT::ExtDataT` in order to create its own "sibling" class hierarchy of windows that parallels the hierarchy in the GuiSys, effectively augmenting the original class hierarchy with any data or functions required for Gui Editor purposes.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@367 5fe0b3f3-48df-4727-b460-e6dda9226989