Model code: `AnimPoseT` now compares `AnimExpressionT` instances ''by value'' in order to determine if it must refresh its mesh caches.
That is, if `Pose` is an `AnimPoseT` instance and `Model` is the related model instance, then in the call sequence
{{{
#!cpp
    Pose.SetAnimExpr(new AnimExprStandardT(Model, -1, 0.0f));
    Pose.Draw(0, 0.0f);

    Pose.SetAnimExpr(new AnimExprStandardT(Model, -1, 0.0f));
    Pose.Draw(0, 0.0f);
}}}
the second call to `SetAnimExpr()` will now no longer trigger cache refreshes in `Pose` (even though the second anim expression is an entirely different instance than the first).

This feature is very important for the worldly use of `AnimPoseT`s and `AnimExpressionT`s, where poses may be shared across different users and the above call sequence occurs frequently.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@459 5fe0b3f3-48df-4727-b460-e6dda9226989