Model code: Added implementations for the tangent-space methods.
Continuing r475, this change adds implementations for the previously added tangent-space computation methods.

Implemented, however, are only the methods `HARD` and `GLOBAL`.
Method `SM_GROUPS` is currently the same as `GLOBAL`, and subject to implementation in separate ticket #101.
Implementing the originally planned t-s method `SG_LOCAL` is unfortunately not worthwhile, as it seems to work well only for cone-shaped meshes (the relevant code is still present, but `#if 0...`'ed).

For #100, it remains to remove the `gts_` ("given tangent-space") code as outlined in the ticket description.

References #100.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@476 5fe0b3f3-48df-4727-b460-e6dda9226989