Net code: Properly deal with `SC1_EntityRemove` messages: Cannot remove an entity from the delta frame that wasn't in the delta frame in the first place.
This was a *very* old and long outstanding bug that fortunately never manifested itself over the years.
It occurs however when a complete `SC1_FrameInfo` message gets too large, so that the server, as an emergency measure, drops some of the related `SC1_EntityUpdate` sub-messages.
When the client then receives the *next* `SC1_FrameInfo` message, a related `SC1_EntityRemove` sub-message may refer to an entity that is not in the clients corresponding delta frame, possibly applying all subsequent delta updates to the wrong entity.

`SC1_FrameInfo` message normally don't get large enough to trigger this, but in the first step of our implementation of #113 we use uncompressed messages that easily start the above chain of events.

This change fixes the problem.

References #113.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@537 5fe0b3f3-48df-4727-b460-e6dda9226989