This change completes the work begun in r538, implementing RLE-compression for the delta messages.
There is a test built in that compares our RLE-compression implementation with the DEFLATE compression algorithm.
When activated, the written `compress_log.txt` file typically looks like this:
{{{
227 bytes in original delta message
33 bytes in deflate-compressed message, compression result is 0 (Z_OK)
28 bytes in RLE-compressed message.

227 bytes in original delta message
12 bytes in deflate-compressed message, compression result is 0 (Z_OK)
6 bytes in RLE-compressed message.

227 bytes in original delta message
12 bytes in deflate-compressed message, compression result is 0 (Z_OK)
6 bytes in RLE-compressed message.

227 bytes in original delta message
32 bytes in deflate-compressed message, compression result is 0 (Z_OK)
28 bytes in RLE-compressed message.

227 bytes in original delta message
34 bytes in deflate-compressed message, compression result is 0 (Z_OK)
27 bytes in RLE-compressed message.
}}}
In our (limited) tests, not a single case occurred where the RLE-compressed message was larger than the deflate-compressed one.

This change, together with r539, remedies the caveats mentioned in r538: Cafu should be good for public use again (even though there is more work to be done for #113).

References #113.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@540 5fe0b3f3-48df-4727-b460-e6dda9226989