Scripting: Provide a Lua method GetRefCount() for all bound C++ objects.
The method returns `nil` for all objects for which a reference count is not applicable (objects that were copied to bind them to Lua),
or the true reference count for all objects that were bound via an `IntrusivePtrT`.

A reference count of 1 means that the object is currently only used in Lua, but nowhere in C++ code.
A reference count of 2 and higher means that there are also references held in C++ code.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@578 5fe0b3f3-48df-4727-b460-e6dda9226989