GuiSys, Scripting: Changed the GuiSys to keep all `WindowT` instances in `IntrusivePtrT`s, and accordingly changed the related scripting implementation to use the `cf::ScriptBinderT` class for binding `WindowT` instances to Lua.
This gets us rid of the GuiSys's prior, proprietary binding implementation that was centered around the (now deleted) `WindowPtrT` class for "anchoring" and "unanchoring" the related Lua instances.
It also means that the ambiguity regarding the parent and list of children is now removed: The C++ `WindowT` code is now the only and thus authoritative place where this information is kept.

The resulting code is shorter, easier to understand, more natural, more flexible and more general.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@584 5fe0b3f3-48df-4727-b460-e6dda9226989