Engine core: Removed member `m_PredictedState` from class `EngineEntityT`.
The member `m_PredictedState` is no longer needed, because the last state as it was sent from the server is now (since r606) kept in the `m_OldStates`. In the state of member `Entity`, we keep the extrapolated states of NPCs.

This revision extends the concept also to human player entities: Their "extrapolation" is really the prediction, which we now keep in the `Entity` itself, too.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@615 5fe0b3f3-48df-4727-b460-e6dda9226989