Game code, engine core: Added an extrapolation "framework" to class `BaseEntityT`. This is the successor of the interpolation code removed in r620.
Its key features are:
  - Better code design: No longer bothers the engine core with interpolation, moved it all into the game code.
  - It is more flexible in that it can extrapolate not only entity origins, but any arbitrary values.

Note that the new extrapolation code is nowhere employed yet!
This is left for the subsequent revision.

Personal note:
The extrapolation framework is encapsulated relatively well in the `BaseEntityT` class.
I wished to have it encapsulated / separated / kept from everything else even more cleanly, but unfortunately have not found any better way to do so.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@621 5fe0b3f3-48df-4727-b460-e6dda9226989