Engine core: The client and server code used to contain hard-coded references / strings to game "DeathMatch". This obviously causes problems for custom games with different names.
To overcome these problems, this change introduces the new `GameInfoI` interface.

The key idea is to build a list of `GameInfoI`s of all games known (linked) to the executable.
This is done early in application init, so that the list can be queried at the command-line ("--help"),
and a game from it be elected (possibly also at the command-line) for use.

Once the choice is made, the related `GameInfoI` is queried for all subsequent initialization.

References #116.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@640 5fe0b3f3-48df-4727-b460-e6dda9226989