Game code, scripting: break "Ent -> GUI -> Ent" cycles.
If an entity holds a GUI, it can be worthwhile to bind the entity to its own GUI, so that the GUI script has convenient access to "its" entity.

This however creates a cycle (Ent -> GUI -> ScriptState -> Ent) that prevents the entity from being destructed when the core engine drops its last `IntrusivePtrT<>` to it (the dtor is never run), which in turn - besides leaking memory - can cause consecutive problems (such as the entities clip model not being removed from the clip world).

Overcoming this problem is however a lot harder than it seems:
  - Binding the entity to the GUI "by raw pointer" is not viable as explained in r648.
  - By their nature, there is no concept of "weak" intrusive pointers, and anyways they would suffer from the same problem as raw pointers (cannot bind the same type multiply with different "binding semantics").
  - Monitoring (via call-back) when the reference count drops to 1 (meaning that only the cyclic reference is left) *might* work. Initially this appears to be elegant and appealing, but on second thought turns out to be less generic and more confusing than expected, and in fact neither simpler nor more elegant than destroying the cycles explicitly.
  - The simplest solution is to explicitly destroy the cycle before the entity is removed from the map.

The last solution is not elegant, but simple and clear, and was thus implemented in this change.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@649 5fe0b3f3-48df-4727-b460-e6dda9226989