Game code: Add `RunMapCommand()` script method to `EntStaticDetailModelT` entities.
GUI scripts that want to run a command in the map script state of the entity that they're attached to no longer call `gui:RunMapCmd()` but `Entity:RunMapCommand()`.
This liberates the GUI from having to know the map script state. Instead, and much more naturally, it is now the entity object (that is bound to the GUI script) that has the required knowledge and implementation.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@653 5fe0b3f3-48df-4727-b460-e6dda9226989