Build system: "Abstract" the game name even in the games SConscript files.
That is, game DeathMatch no longer names its own library "DeathMatch",
but according to the `GameLib` variable that it is passed from the SConstruct file.

This change is an important prerequisite so that new games, that are started as a
plain copy of an existing game, can automatically be compiled and linked.



git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@658 5fe0b3f3-48df-4727-b460-e6dda9226989