GuiSys, CaWE: Improve window naming.
Class `cf::GuiSys::WindowT` now enforces that the name of a window is unique among its siblings at all times, modifying it as necessary.
The effects of this on the GUI Editor eventually fix #124.

Thanks to Tyson for the original [http://www.cafu.de/forum/viewtopic.php?p=5425#p5425 report]!


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@667 5fe0b3f3-48df-4727-b460-e6dda9226989