GUI Editor: Clear clipboard in time during shutdown.
The GUI Editor clipboard may still contain objects that in turn have resources e.g. in the MatSys when the application shuts down.
This change moves the GUI Editors clipboard into the `ParentFrameT` and makes sure that during shutdown, the clipboard is cleared in time.

The same problem may possibly still exist with the Map Editor, see #9.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@670 5fe0b3f3-48df-4727-b460-e6dda9226989