Server: Loudly complain if type info for C++ entity class is not found.
Before this change, when a developer added a new C++ entity class and forgot to update the `AllTypeInfos[]` list in `GameImpl.cpp`, only a warning message was printed to the console that did not easily catch the attention of the developer.
We now show an explicit dialog box that prominently indicates the problem and suggests the likely solution.

Fixes #123.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@682 5fe0b3f3-48df-4727-b460-e6dda9226989