GuiSys: Update `ComponentBaseT::UpdateDependencies()` method.
It becomes clear that `ComponentBaseT::ResolveDependencies()` is called not only when a window and its components have been loaded (a one time issue at init time), but also when a window is copied or when it has been modified in the GUI editor.
This change reflects this insight by renaming the method and updating its documentation.

References #125.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@687 5fe0b3f3-48df-4727-b460-e6dda9226989