CaWE: Introduce component system to GUI Editor.
The recently added component system for GUI windows is now also introduced to the GUI Editor in CaWE:
The GUI Editor gets a "Component" menu and a "add component" command so that the user can add components to windows.
(At this time however, the "Window Inspector" has not yet been updated, so that the newly added components are not yet visible in the user interface.)

References #125.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@690 5fe0b3f3-48df-4727-b460-e6dda9226989