GUI Editor: In `WindowInspectorT`, use `EVT_PG_CHANGING` for component properties.
We used to handle changes of our component properties along with the normal (somewhat outdated) window properties in `WindowInspectorT::OnPropertyGridChanged()`, where however it is too late to veto the related event.
As the window properties rely on the "changed" (not the "changing") event, we separated our component properties from them, which are now handled in the new `WindowInspectorT::OnPropertyGridChanging()` handler.

References #125.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@696 5fe0b3f3-48df-4727-b460-e6dda9226989