GuiSys: Add support for getting/setting component variables in Lua.
The values of component variables can now be retrieved and set in Lua.
If `comp` is a component in Lua, and `filename` is a variable of `comp`, then we can get and set its value like this:
{{{
#!lua
    comp:set("filename", "image.png")
    print(comp:get("filename"))
}}}


Note that at this time, there is no support yet for creating new components in Lua,
and there is no support yet for the more natural short form:
{{{
#!lua
    comp.filename = "image.png"
    print(comp.filename)
}}}

References #125.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@698 5fe0b3f3-48df-4727-b460-e6dda9226989