GuiSys: Allow components to exist "stand-alone": independent and not a part of any window.
Components now have two clearly defined invariants: "part of" and "not part of" a window.

The normal invariant is that a window knows the components it is composed of,
because it hosts them, and reversely, that each component knows its parent
window that hosts it.

In the second invariant, a component is independent of any window: neither is it part
of the components list of a window, nor does it reversely have a parent window assigned.

The latter is very useful when windows are dynamically constructed, or components are
added to or removed from windows at runtime. This occurs in the Undo/Redo system of our
GUI Editor, as well as when components are created and handled in scripts.

Also see r685, which did not foresee that true "stand-alone" would be required.
References #125.


git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@699 5fe0b3f3-48df-4727-b460-e6dda9226989