Map Editor (groups toolbar): Replace obsolete icons.
The Groups toolbar still used some old icons, for which since long we have
better ones in CaWE/res/icons/, so use those.

Also it seems that if we use menu icons directly from the `wxArtProvider`
instead of the image list, the alpha transluceny is properly handled also
under wxMSW (see previous revision for a description).