CaWE: Made "Silk" icons accessible to the `wxArtProvider`.
Moved the icons from the famfamfam.com "Silk" icon set into a common directory
that is accessible to the `wxArtProvider`.

With the icons for the menu items now coming directly from the `wxArtProvider`
rather than the image list, the rendering problems mentioned in a58a6e5 no
longer occur / are fixed.