GUI Editor: Put the first component into effective operation!
This change augments the `ComponentImageT` so that it is capable of doing what
the "Background" properties of the `WindowT` class are doing.

The "image" component can now be used as a full replacement of the "Background"
properties. Both features will co-exist for a short while, until we switch all
existing `.cgui` files to use the "image" component and eventually remove the
"Background" properties from the `WindowT` class.

References #125.