TypeSys: Add support for "extra messages" to `VarBaseT`s.
As the comment for `VarBaseT::GetExtraMessage()` already indicates,
this change is probably not the last word on the subject: It's more
a work-around to communicate model loading failure to the user.

Better solutions, such as a concise logging system, are not easily
implemented though, so for now the `GetExtraMessage()` mechanism
seems to be a reasonable compromise.