TypeSys: Add support for "choices" to `cf::TypeSys::VarT<T>` variables.
Choices can dynamically be determined at run-time, as is important whenever
the set depends on an external factor. For example, the available animation
sequences or skin numbers change whenever the related model changes.

Right now, although choices are generic and work with all types `T`,
the GUI Editor only accounts for them if `T` is `int`.
It is easy however to extend this support to cover the other types as well.