GUI Editor: Put the second component `ComponentModelT` into effective operation!
This change augments the `ComponentModelT` so that it is capable of doing what
the "Model..." properties of the `ModelWindowT` class are doing.
As indicated in the TODO comment in the code, there are still a few things that
need to be done, but the bulk of the code, a good basis, is there.

References #125.