GUI Editor, Window Inspector: When a variable changes, update the whole component.
When the value of a component variable changes, the side-effects of the change
may have changed other variables in the same component ("siblings") as well.

Therefore, in the Window Inspector, when we receive a notification about the
change of a component variable, update the grid properties of *all* variables
in the component.

In addition to that, when a property is updated whose associated component
variable has "choices", the set of choices itself is updated before the new
value of the variable is set for the property.

This, together with the previous revision c02ceda, fixes and completes the
handling of component variables in the GUI Editor for which the side-effects
of setting a new value extend to other variables (in the same component) as
well.

References #125.