GuiSys: When a `ComponentModelT`s "Animation" (sequence number) variable is set, have it take effect immediately, don't blend.
Setting a new sequence number for the model of a `ComponentModelT` used to add a "blend" anim expression to the models pose for blending from the previous to the new animation sequence. This however looks wrong in the GUI Editor, where no `OnClockTickEvent()` methods for advancing the time are implemented yet, so that the blend cannot proceed. With the blend being stuck at "frame 0", it seems as if setting a new animation had no effect at all.

Moreover, implying a "blend" where a "set" was requested seems semantically wrong:
Added an explicit script function for blending to a new animation sequence instead.

References #125.