Engine: Provide the game code access to the commonly used `GuiResourcesT` instance.
If a game entity wants to create its own "private" GUI (without using the
global `GuiManI` interface), it must know the proper `GuiResourcesT` instance.

This change provides the `ClientT`, `ServerT` and `Ca3DEWorldT` classes each
with a reference to the commonly used `GuiResourcesT` instance, so that the
game (entity) code can obtain it at any time via the new
`GameWorldI::GetGuiResources()` method.

Note that this change adds 1:1 analogously for the `GuiResourcesT` what was
already there for the equally commonly used `ModelManagerT`.