Game code: Create stand-alone instances of 3D world GUIs, no longer manage them in the `GuiMan`.
Registering 3D world GUIs with the global `cf::GuiSys::GuiMan` used to be
necessary when the game code was still in DLL form, dynamically linked.
However, doing so implied the problem described in #23 in the first place.
Now that we no longer link dynamically and have the common `GuiResourcesT`
readily available, it is much more natural to have each `EntStaticDetailModelT`
have its own, independent GUI instance.

Closes #23.