GUI Editor: Convert subclasses of `WindowT` to components when a GUI file is saved.
This is intentionally limited to subclasses of `WindowT` for now. That is,
conversion of `WindowT` details itself (e.g. text, border, image) is currently
omitted.

This separation into two steps improves the clarity of the conversion, and
gives us a better chance to fix custom GUI scripts as an intermediate step.

References #125.