DeathMatch GUIs: Manually split old `.cgui` files into `_init.cgui` and `_main.cgui`.
The files addressed here were hand-written at a time before the GUI Editor was available.
It was possible to load them into the GUI Editor, but not to save them back without loss.
The purpose of this change is to make this possible, but also to maintain
backwards-compatibility with existing maps and code, which continue to refer
to the GUIs under their old filenames.