DeathMatch GUIs: Apply one load-save cycle to each formerly hand-written GUI.
As prepared in the previous revision, our old, formerly hand-written GUIs can
now be loaded and saved again by the GUI Editor without loss.
This change is exactly that: the result of one load-save cycle for each
formerly hand-written GUI, plus manual fixes for any encountered problems.

The goal is to "streamline" these GUIs, bringing them on par with the rest,
so that the upcoming conversion to the component system becomes easier.