GuiSys: Change signature of `ComponentListBoxT::OnSelectionChanged()` callback, remove `SelNum` parameter.
We used to pass the index number of the newly selected item before, but it has
been removed and the signature is now empty.

The main reason for this change was that the index number, e.g. named `SelNum`,
was in `OnSelectionChanged()` implementation often used like this:

    local Item = self:get("Items")[SelNum]

However, this is now much better written as

    local Item = self:GetSelItem()

Note that the code can still easily learn the same `SelNum` as before:

    local SelNum = self:get("Selection")

References #125.