GuiSys: Add script method `gui:Init()`.
With the new method, GUI scripts can have the `OnInit()` window callbacks be
processed before it would automatically be done at the end of the main script.

This is useful whenever the main script wants to rely on the effects of the
`OnInit()` methods, e.g. the proper default values all being set, the
components being instantiated, etc.

References #125.