DeathMatch GUIs: (partially) convert GUIs to component system, covering `WindowT` attributes.
This change continues the conversion begun in [changeset:41adfa/git] of 2013-02-25.

Using the code introduced in [changeset:caec74/git] on 2013-03-08, continue converting the GUIs to the new component system, now covering the attributes of class `WindowT`.

This was done by loading and saving all `_init.cgui` files ''twice'':
  - first to have them re-written by the conversion code,
  - then to have them cleanly written by the normal save code.

The conversion is partial, because:
  - Now also the attributes of `WindowT` are converted, but the window rectangles are still written in the old way, too, in order to ease the upcoming transition in the code.
  - It doesn't work: The scripts in the related `_main.cgui` files must now be fixed, especially interpolations.

References #125.