GUI Editor: Properly handle `cf::TypeSys::VarT<>`s that "clean" values in their `Set()` method.
When the user enters a new value in the "Window Inspector", eventually code like
{{{
#!cpp
    Var.Set(Event.GetValue());
}}}
is called. Note that if `Var` is of a type derived from `cf::TypeSys::VarT<>`, the `Set()` method may well set `Var` to a value that is different from `Event.GetValue()`.

This change makes sure that the property in the "Window Inspector" reflects the actual value of `Var`.

References #125.