GuiSys: Flag the "Basics" and "Transform" components as "fundamental".
Fundamental components are explicit C++ members of class `WindowT`.
Users cannot delete them in the GUI Editor or via script, and cannot add new
ones in the GUI Editor. (It is possible to add additional ones via script,
but not useful.)
Note that most components are not fundamental, but "custom":
Users can freely add or delete them in the GUI Editor or via scripts.

References #125.