DeathMatch GUIs: Revise GUIs, fix the remaining conversion artifacts.
After the multi-step conversion of the GUIs to the new component system, the components were converted properly, but not always in the order that produced the same visual result as before.
More precisely, some "Image" components that were intended as background images were too early in the order and thus rendered too late, so that they covered the background and unintentionally the text as well. This change brings such components in proper order.

Also, the old system did not have a "padding" feature for the rendered text, and had to compensate this lack with cleverly inserted space characters or offset window positions. This change also removes the old compensations and sets proper "padding" instead.

References #125.