GUI Editor: No longer clone the entire window hierarchy in the `GuiDocumentT` ctor.
In the past, we had a reason for not directly working with the native window hierarchy that was kept in the `m_Gui` instance, but to clone it and to only work with the cloned hierarchy.

Unfortunately, I cannot remember what that reason was, and repository archeology did not reveal it. However, I believe that it was related to the way in which we kept C++ `WindowT` instances and their Lua "alter egos" in sync, before the binding code was modernized to use reference-counted pointers.

In any case, whatever the reason was, it no longer seems to be valid, and as cloning the window hierarchy only seems to be an extra effort that helps nothing but confusion, it was finally removed.