GuiSys: Augment `WindowT` so that it can hold an application-specific component.
The `WindowT` holds the new application component very much like the "Basics" and "Transform" components, but knows next to nothing about it (except that it is derived from `ComponentBaseT`).
The application component can be set by the application, and is in fact intended for the sole use by the application. For example, the GUI Editor could use it to implement a "selection gizmo".

References #125.