GuiSys: Moved all the TypeSys-related definitions to the bottom of each source file.
This changes prepares an upcoming enhancement to the TypeSys (adding
documentation strings for the methods of a type) that will make it favorable
to keep the TypeSys-related definitions to the bottom of each source file.

Besides that, when reading the source code the TypeSys-related definitions
are less intrusive when they're at the bottom of the file.