TypeSys: Augment `TypeInfoT`s to keep optional documentation strings about the type.
The class itself, its methods and its variables can be augmented with
documentation strings. These are intended to store more documentation about a
type directly in C++ code, so that auto-generating scripting documentation
becomes easier and more useful.

Note that the `DocMethods` correlate to the `MethodsList`, whereas the
`DocVars` have no such correspondence.