Map Editor: Save the `m_ScriptWorldT` contents when saving the map.
The contents of the `cf::GameSys::WorldT` is saved to a new `.cent` file
next to the usual `.cmap` file.

Later, I expect it to be loaded with both the `.cmap` (in the Map Editor)
as well as with the `.cw` files (in the Cafu Engine).

Also upgrade the `.cmap` file format version to 14, because beginning with
version 14, we will expect the related `.cent` file to exist when the `.cmap`
file is loaded.