Map Editor: Add new map primitive `MapEntRepresT`.
The new map primitive `MapEntRepresT` provides a graphical representation of
an entity in the Map Editor.

It is supposed to gather the Map Editor specific code of the `MapEntityBaseT`
classes, separating entity-specific code from Map Editor specific code more
clearly.

The goal is to no longer derive the `MapEntityBaseT` class from `MapElementT`,
so that it becomes easier to change the `MapEntityBaseT` entity class(es).