Map Editor: Derive `MapEntRepresT` from `MapElementT` rather than from `MapPrimitiveT`.
This change clearly separates the `MapEntRepresT` type from those derived from
`MapPrimitiveT`. This way it becomes obvious and easy to distinguish between:

   - `ArrayT<MapElementT*>` -- can also contain `MapEntRepresT`s
   - `ArrayT<MapPrimitiveT*>` -- cannot contain `MapEntRepresT`s

and the C++ type system can help us to not mistakenly insert a `MapEntRepresT`
into an array where only `MapPrimitiveT`s are possible (e.g. the `m_Primitives`
member of `MapEntityBaseT`).

In this sense, the clear distinction between `MapElementT`s and
`MapPrimitiveT`s is also a great help in documentation and communication.