Map Editor: Revise the cloning of `MapEntRepresT`s.
What used to be a complicated mutual relationship before has now been replaced
by a much simpler setup, where the cloned `MapEntRepresT` instance actually
holds the cloned entity.

This amends the previous revision, as it removes the `MapEntRepresT`'s friend
access to the `MapEntityBaseT`, which as per the previous revision is to be
fully encapsulated by the `CompMapEntityT` class only.