Map Editor: Fix `MapDocAdapterT` usage.
The code used to pass temporary instances of `MapDocAdapterT` to the
`MaterialBrowser::DialogT` ctor, which keeps a reference to the passed instance
for later use, e.g. while the dialog is shown with `ShowModal()` or even
thereafter.

The fact that this worked is more luck than anything else, because the
`MapDocAdapterT` instance is actually gone after the ctor completes.

One way to fix this was to create the `MapDocAdapterT` instance explicitly
outside of and before the dialog, e.g.

    MapDocAdapterT           DocAdapter(m_MapDoc);
    MaterialBrowser::DialogT MatBrowser(this, DocAdapter, ...);

but as we need longer-lived `MapDocAdapterT` instances for the properties of
the upcoming Entity Inspector soon anyway, we instead add an instance into the
`MapDocumentT` itself (analogous to the `GuiDocumentT` class) and use that.