GameSys: Update the classes so that components (especially "Model") can be rendered in user code.
Especially the "Model" component was still very much like the GuiSys's "Model"
component that it was originally copied from. This change makes the appropriate
updates and adds support for GUIs that can be attached to the models GUI
fixtures.

With these changes, the Map Editor can actually use the entity components in
its 3D view renderings.