Let all world classes have a reference to a `cf::GuiSys::GuiResourcesT` instance.
The reference to the `cf::GuiSys::GuiResourcesT` instance is now propagated
through all related world classes, in the same manner as it is already done
for the analogous, "parallel" `ModelManagerT` resource.

With this change, we're prepared to introduce `cf::GameSys::WorldT` instances
as part of the classic `WorldT` classes soon.