Server: Create `cf::GameSys::EntityT` instances for newly created game entities.
As a side effect, this change also updates the signature of the
`CaServerWorldT::CreateNewEntityFromBasicInfo()` method.

Note that up to this point, the Cafu Engine itself (the client, the server) is
still not aware of and knows nothing about the new `cf::GameSys::EntityT`s --
the `EngineEntityT`s have no pointer or reference to them.