Common: Fix the design issues that were introduced several revisions ago.
It was an error to add a `cf::GameSys::WorldT` as a member to the common
`WorldT` class in revision 59f6f5dac of 2013-06-29 and to build on it in
revision b9f1ae1c0 of 2013-07-05.

When the Cafu Engine runs in single-player or server-with-local-player mode,
a single instance of the common `WorldT` class is shared by the Cafu client
and the server.
Consequently, the `WorldT` class can only have `const` state that is common
to both, and not anything dynamic as we erroneously had before with the
`cf::GameSys::WorldT` instances.