Engine: Create an `EngineEntityT` instance also for the world (root) entity.
The purpose of this change is to create the basis for a 1:1 mapping of
`EngineEntityT`s to `cf::GameSys::EntityT`s.

No bad side effects are supposed to occur from this change:
All code (client, server, game, network messages) should behave as expected.
A network dump that I analyzed with Wireshark (and compared to a previous one)
looked as expected as well.