cw files: Add methods to undo the scaling by 25.4 that was applied when the world was initially loaded from the cmap file, getting back from millimeters to "world units".
This change is intended to provide the Cafu Engine with worlds that are in the
same unit as the maps in the Map Editor.

Of course, this change is a detour: Instead of removing the upscaling in the
first place, we left CaBSP, CaPVS and CaLight in mm, whereas the Cafu Engine
calls the new extra methods to scale down again.

This was done because removing the scaling triggers rounding errors, and even
with updated epsilon values, we found that CaBSP then produces vastly different
results than before. Removing the scaling is thus more work than it initially
seems, and this - temporary - change allows us to defer it until later.