Game code: Convert all units from millimeters to "world units".
The goal of this commit series is to drop millimeters and instead to work with
"world units" (the same unit as in the Map Editor) everywhere.

As a result, the user, the core engine and the game code will never see
millimeters, but only know and work with "world units".

The map compile tools CaBSP, CaPVS, CaLight and CaSHL are, for now, an
exception: Changing them from millimeters to "world units" can easily cause
new problems due to rounding errors, whose careful investigation we defer
until later.