GameSys: Revise components to properly deal with unchanged values.
The deserialization of network messages on the client can cause a member
variable of a component to be `Set()` very frequently, and often to the same
value as before.

Consequently, setting a variable to the same value must be dealt with as
efficiently as possible (for performance), and free of unwanted side effects
(for correctness).

As an example, in the `ComponentModelT` setting the animation sequence number
to the same value as before must neither re-instantiate an anim expression
(for performance), nor reset the existing anim expression to frame 0 (for
correctness, or else the model cannot properly play the animation sequence).

As other examples, setting the model or GUI names to the same values as before
must ''not'' cause the related instances to be re-initialized or
re-instantiated (for performance).