Game code: Provide the "old-style" `BaseEntityT`-derived entities with a pointer to their "new-style" `cf::GameSys::EntityT` counterpart.
This helps with transferring more and more functionality into the new GameSys
code smoothly and in small steps.

That is, functionality that has been (re-)implemented at the
`cf::GameSys::EntityT` or `cf::GameSys::Component*T` level can immediately be
(re-)used for replacing the implemenation in the `BaseEntityT` entities.
This in turn allows smaller steps and for easier and more complete testing in
the progress towards moving *all* functionality into the new entity code.