Math: Complete the round-trip: `AnglesT <--> Matrix3x3T <--> QuaternionT <--> AnglesT`
Each of the three can now be constructed from each of the two others.
This very important step finally gives us the completeness and consistency
that we have always wanted and needed for representing orientations.

The new code will allow us to replace a few older, almost-but-not-quite
equivalent methods such as `QuaternionT::Euler()` and `Matrix3x3T::*_COMPAT()`
in the future.