Scripting: Augment `UniScriptStateT::CallMethod()` to accept extra arguments that the user has pushed on the stack.
The key idea is to overcome the limitations of the "..." vararg mechanism.
With this commit, the `CallMethod()` seems flexible enough, but is certainly
still from ideal.
Implementing this with variadic templates seems like a good option, but also
seems to require us to drop support for all pre-2013 versions of Visual C++.
Yet other alternatives that use e.g. an inheritance hierarchy seem possible,
too, but are most likely not worth the effort.