Engine, GameSys: Don't process events if we're delta'ing from the baseline (e.g. when an entity re-enters the PVS).
Before this change, it was possible that we walked over a weapon to pick it up,
quickly left and turned around a corner or two, and lost the weapon from our PVS
before it had a chance to respawn.
If we later moved back, so that the weapon re-entered the PVS, it would
erroneously trigger the re-spawn event (e.g. play the appropriate sound), even
though the real event was long over.