Introduce end-of-line normalization (same as in the work branches).
See the "gitattributes" man page for the renormalization recipe:

    $ rm .git/index     # Remove the index to force git to
    $ git reset         # re-scan the working directory
    $ git status        # Show files that will be normalized

Only few files were actually affected by the renormalization step, thanks to
our prior careful attention to check-in vendor libraries with LF line endings
in the first place.

Note that since we migrated the Cafu repository from Subversion to Git, only
the Lua library was upgraded and merged into one of the development branches.
Doing this worked without problems regarding the EOL style only because we
happened to check-in Lua with LF line endings. Had we committed the new Lua
with CRLF line endings instead, there would have been merge conflicts.
It would have been possible to set the `merge.renormalize` configuration
variable to true, but it only seems to be right to use the same .gitattributes
file in the vendor branch as in the work branches.