Fix all issues identified by Visual C++ 2012 Express Edition's `/analyze` option (that were not false positives).
Unfortunately, the `/analyze` feature produces relatively many false positives
that I had to squelch.

There were not many real issues left (certainly much fewer than I expected),
most of which were of type

    warning C6246: Local declaration of <variable> hides declaration of same name in outer scope.

and none of which were serious.

As an early conclusion, our code base seems to be in a pretty good shape -- and
we should probably repeat this check with the full or a newer version of Visual
C++ 2012, or another static code analysis tool that is more aggressive.