GuiSys: Separate the GUI's script state from the `GuiImplT` instance.
`GuiImplT` instances should not contain their own script state instances, but
script states should exist first, and GUIs be added (bound) to them.
This increases the flexibility of both GUI as well as script state instances.

Note that in change cc5cad4 of 2013-11-03, we did the same change analogously
for `cf::GameSys::WorldT`s. However, as opposed to worlds, with GUIs there is
an extra twist for backwards-compatibility: At this time, separating the GUI's
script state from the `GuiImplT` instance is *optional*, so that user code that
is difficult to update can be addressed at a later, more convenient time.