DeathMatch: Re-implement the old `EntFuncMoverT` C++ code in terms of the entity component-system, employing a script component.
This change also ports the old test/demo script code for the TestPatches map:
All code that used to be in `Games/DeathMatch/Worlds/TestPatches.lua` is now in
script components of the entities in the TestPatches map.

Not yet implemented is the pushing of other entities that are contacted while
the mover is moving.