GameSys: Introduce `OnSequenceWrap()` as a new callback method to Model components.
Model components call the user-defined script method `OnSequenceWrap()`
whenever their current animation sequence "wraps".

More precisely, if the animation expression of the pose of a Model component is
a "standard" expression (as opposed to e.g. a blending or filtering expression)
that is defined as "looping", and the sequence of that anim expression
"wraps", the `OnSequenceWrap()` script method is called.