Doxygen: Support the documentation of "overloads" for scripting methods.
It may be desired to use "overloads" when documenting scripting methods,
whenever the implementation of the scripting method is able to take a varying
set of parameters so that it looks as if "true" overloading as in C++ was used.

This change also adds two small fixes regarding the automatic generation of the
Doxygen template header files.