GameSys: Add virtual method `GetCollisionBB()` to the `ComponentBaseT` hierarchy.
This method is necessary so that any component can contribute to the collision
shape that is eventually used for the rigid body for physics computations
(especially the `ComponentCollisionModelT` with external collision models and
`CompGameEntityT` with the map primitives),
and so that any (other) component can learn this information, especially the
(upcoming) `ComponentPhysicsT` physics component for rigid bodies that are
dynamically simulated under the laws of physics, and (later) the
`CompGameEntityT` application component for the rest, i.e. bodies that are
entirely static or kinematic only, but must still be a part in the physics
world.

Note that the "restriction" to a bounding-box is purely coincidental: It was
chosen because it makes the implementation of the (upcoming)
`ComponentPhysicsT` component simple.
It should later be replaced by something that is more accurate, e.g. a convex
hull, the list of actual collision models, or something similarly useful.