GameSys: Add new component `ComponentPhysicsT`, and re-implement "rigid body" entities in its terms.
This component is only about "dynamic" physics objects, and the handling of
physics bodies that are static (don't move at all) or kinematic (moved by script
or program code) must still be added (to the `CompGameEntityT` application
component).

Nevertheless, this is a very nice milestone in the transition to the new
component system!   :-)